+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 0.48.0 - 2026-01-27
+
+### Enhancements
+- Added initial support for splitting DBN files:
+  - Added new `SplitEncoder` that supports synchronous and asynchronous encoding
+  - Added new `Splitter` trait that allows for extensible splitting of files while
+    reusing the `SplitEncoder` boilerplate.
+  - Added `SchemaSplitter`, `SymbolSplitter`, and `TimeSplitter` which allow for
+    different methods of splitting DBN files
+  - Added split support to the CLI. For example:
+    `dbn mbo.dbn --split-by week --output-pattern '{date}.json'` --json
+- Added new publisher for Blue Ocean ATS (`OCEA_MEMOIR_OCEA`)
+
+### Bug fixes
+- Fixed issue where `AsyncDynReader` instances created from `with_buffer()` would only
+  decode the first frame of multi-frame Zstandard files
+
 ## 0.47.0 - 2026-01-20
 
 ### Enhancements